Passage des modales Theme et Consentement en composant
